change docker to systemd

+++ scripts/setup.sh
@@ -1,13 +1,56 @@
 #!/bin/bash
 
-ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )/.."
+# Check if running as root
+if [ "$(id -u)" -ne 0 ]; then
+    echo "This script must be run as root"
+    exit 1
+fi
+
+# Stop running Docker Compose containers
+if type docker >/dev/null 2>&1; then
+    docker compose down
+fi
+
+# Detect basic variables
+ROOT=$(realpath "$(dirname "${BASH_SOURCE[0]}")/..")
+SERVICE_NAME=$(basename "$ROOT")
 
 # Configure Git
 git config pull.rebase false
 
-# Create Config File
+# Validate the binary file
+BINARY_PATH="$ROOT/p-node"
+if [ ! -f "$BINARY_PATH" ]; then
+    echo "Binary not found at $BINARY_PATH"
+    exit 1
+fi
+
+# Create the config file
 if [ ! -f "$ROOT"/configs/main.json ]; then
 		cp "$ROOT"/configs/main.defaults.json "$ROOT"/configs/main.json;
+fi
+
+# Setup Systemd
+SERVICE_FILE="/etc/systemd/system/$SERVICE_NAME.service"
+SERVICE_TEMPLATE="$ROOT/scripts/service.template"
+
+sed "s|THE_NAME|$SERVICE_NAME|" "$SERVICE_TEMPLATE" > "$SERVICE_FILE"
+sed -i "s|THE_PATH|$BINARY_PATH|" "$SERVICE_FILE"
+sed -i "s|THE_DIR|$ROOT|" "$SERVICE_FILE"
+systemctl daemon-reload
+
+if systemctl is-enabled --quiet "$SERVICE_NAME"; then
+    echo "Service $SERVICE_NAME is already enabled."
+    echo "Restarting service $SERVICE_NAME..."
+    systemctl restart "$SERVICE_NAME"
+    echo "Service $SERVICE_NAME restarted."
+else
+    echo "Enabling service $SERVICE_NAME..."
+    systemctl enable "$SERVICE_NAME"
+    echo "Service $SERVICE_NAME enabled."
+    echo "Starting service $SERVICE_NAME..."
+    systemctl start "$SERVICE_NAME"
+    echo "Service $SERVICE_NAME started."
 fi
 
 # Setup Cron Jobs
